Refactor repo cache into FormCache context
