+++ src/hooks/useFormCache.ts
@@ -4,16 +4,28 @@ import { FormCacheContext, IFormCache } from "../context/FormCache";
 function useFormCache(): IFormCache {
   const {
     getChoiceOptions,
-    cacheChoiceOption
+    cacheChoiceOption,
+    getRepositoryOptions,
+    getRefOptions,
+    cacheRepositorySelection,
   } = useContext(FormCacheContext) as IFormCache;
 
-  return useMemo(() => ({
-    getChoiceOptions,
-    cacheChoiceOption
-  }), [
-    getChoiceOptions,
-    cacheChoiceOption
-  ]);
+  return useMemo(
+    () => ({
+      getChoiceOptions,
+      cacheChoiceOption,
+      getRepositoryOptions,
+      getRefOptions,
+      cacheRepositorySelection,
+    }),
+    [
+      getChoiceOptions,
+      cacheChoiceOption,
+      getRepositoryOptions,
+      getRefOptions,
+      cacheRepositorySelection,
+    ],
+  );
 }
 
 export default useFormCache;
